Fix selectors for checking official ref status
