+++ client/src/js/references/__tests__/utils.test.js
@@ -1,31 +1,6 @@
-import { checkHasOfficialRemote, removeMember } from "../utils";
+import { removeMember } from "../utils";
 
 describe("Reference utils", () => {
-    describe("checkHasOfficialRemote", () => {
-        it("returns false when documents empty", () => {
-            const documents = [];
-            const result = checkHasOfficialRemote({ documents });
-            expect(result).toBe(false);
-        });
-
-        it("returns false when no official refs found", () => {
-            const documents = [{ id: "official" }];
-            const result = checkHasOfficialRemote({ documents });
-            expect(result).toBe(false);
-        });
-
-        it("returns true when official ref found", () => {
-            const documents = [
-                {
-                    id: "official",
-                    remotes_from: { errors: [], slug: "virtool/ref-plant-viruses" }
-                }
-            ];
-            const result = checkHasOfficialRemote({ documents });
-            expect(result).toBe(true);
-        });
-    });
-
     describe("removeMember", () => {
         it("has no effect on empty documents", () => {
             const result = removeMember([], []);
