+++ client/src/js/references/utils.js
@@ -1,8 +1,5 @@
 import { pull, reject, some } from "lodash-es";
 
-export const checkHasOfficialRemote = state =>
-    some(state.documents, ["remotes_from.slug", "virtool/ref-plant-viruses"]);
-
 export const removeMember = (list, pendingRemoves) => {
     const target = pendingRemoves[0];
     pull(pendingRemoves, target);
